requirements.txt
Django==4.2.9
djangorestframework==3.15.0
djoser==2.2.2
phonenumberslite==8.13.32
django-phonenumber-field==7.3.0
pillow==10.2.0